TAP-103: Tidier configurator
